Upgrade Eleventy from 2.0.1 to 3.0.0
Upgrade Eleventy and its RSS and syntax highlighting plugins to their respective latest versions. As part of this, a few breaking changes needed to be addressed. Namely, "rssLastUpdatedDate" was replaced with "getNewestCollectionItemDate | dateToRfc3339" in rss.html. Also, "rssDate" was replaced with "dateToRfc3339" in the same file. Furthermore, "post.templateContent" was replaced with "post.content", also in the same file. One thing that I'm unable to address is the incompatibility of the current Eleventy versions with the eleventy-plugin-pwa package which does lead to the error "Package subpath './src/Eleventy' is not defined by 'exports'" during the Eleventy build process. However, this does not prevent the site from building correctly as far as I can tell. There doesn't seem to be a direct fork of the package that is still maintained so I'm unable to address this break.

Convert the remaining JavaScript files in the project (i.e. the Eleventy config file and related filter and transform scripts) to use ES module syntax to make future updates and migrations simpler. The changes are exclusively a search-and-replace of all remaining "require" calls with their "import" equivalent and of all remaining "module.exports =" with "export default". Also, set the type field in the package.json file to "module".